[wpilibc] Rename SpeedControllerGroupTest.cpp
Also move motor controller test files into motorcontrol folder to match
Java.
