+++ wpilibc/src/test/native/cpp/drive/MecanumDriveTest.cpp
@@ -2,9 +2,9 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-#include "MockMotorController.h"
 #include "frc/drive/MecanumDrive.h"
 #include "gtest/gtest.h"
+#include "motorcontrol/MockMotorController.h"
 
 TEST(MecanumDriveTest, CartesianIK) {
   // Forward
