+++ wpilibc/src/test/native/cpp/motorcontrol/MockMotorController.cpp
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-#include "MockMotorController.h"
+#include "motorcontrol/MockMotorController.h"
 
 using namespace frc;
