+++ wpilibc/src/test/native/include/motorcontrol/MockMotorController.h
